# specs/001-ai-textbook-physical-ai/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Native Textbook API
  description: API for the Physical AI & Humanoid Robotics textbook application
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
paths:
...
  /modules:
    get:
      summary: Get all content modules
      tags:
        - Content
      parameters:
        - name: module_type
          in: query
          schema:
            type: string
            enum: [ros2, gazebo_unity, nvidia_isaac, vla]
          description: Filter by module type
        - name: week_number
          in: query
          schema:
            type: integer
          description: Filter by week number (1-13)
      responses:
        '200':
          description: List of content modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentModuleResponse'
  /modules/{module_id}:
    get:
      summary: Get a specific content module
      tags:
        - Content
      parameters:
        - name: module_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Content module details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentModuleWithContentResponse'
        '404':
          description: Module not found
...
components:
  schemas:
...
    ContentModuleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        module_type:
          type: string
          enum: [ros2, gazebo_unity, nvidia_isaac, vla]
        week_number:
          type: integer
        created_at:
          type: string
          format: date-time
    ContentModuleWithContentResponse:
      allOf:
        - $ref: '#/components/schemas/ContentModuleResponse'
        - type: object
          properties:
            content:
              type: string
              description: MDX content for the module
            prerequisites:
              type: array
              items:
                type: string
                format: uuid
            learning_outcomes:
              type: array
              items:
                type: string
...